requirements.txt
asgiref==3.7.2
bleach==6.0.0
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-recaptcha==3.0.0
idna==3.4
Markdown==3.4.4
Pillow==10.0.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
webencodings==0.5.1
